Refactor auth modules
