+++ pycommerce/infra/api/dependencies/auth.py
@@ -1,15 +1,17 @@
-from uuid import UUID
-from operator import attrgetter
 from datetime import datetime, timedelta
-from typing import Any, Dict, Annotated
-from pydantic import BaseModel
-from jose import JWTError, jwt
+from operator import attrgetter
+from typing import Annotated, Any, Dict
+from uuid import UUID
+
 from fastapi import Depends, HTTPException
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from jose import JWTError, jwt
+from pydantic import BaseModel
+
 from pycommerce.config import get_settings
-from pycommerce.core.entities.user import UserResponse
-from pycommerce.core.services.user import fetch_by_id
-from pycommerce.infra.api.dependencies.repositories import UserRepo
+from pycommerce.core.dtos.user import UserResponse
+from pycommerce.core.usecases.user import get_by_id
+from pycommerce.infra.api.dependencies.user import Repo
 
 _secret_key, _expire_minutes, _algorithm = attrgetter(
     "JWT_SECRET_KEY", "JWT_ACCESS_TOKEN_EXPIRE_MINUTES", "JWT_ALGORITHM"
@@ -50,9 +52,9 @@ def create_access_token(data: Dict[str, Any]) -> Token:
     return Token(access_token=access_token, expire=expire.timestamp())
 
 
-async def get_current_user(repo: UserRepo, token: Oauth2Token) -> UserResponse:
+async def get_current_user(repo: Repo, token: Oauth2Token) -> UserResponse:
     _id = decode_token(token)
-    user = await fetch_by_id(repo, _id)
+    user = await get_by_id(repo, _id)
     if user is None:
         raise CredentialsException
     return user
